Ryddet i noen testfiler
